Feature/#15: KeypadButtonType 코드 위치 변경
