--- Haruby-iOS/Sources/Utility/Constant/CalculatorSymbol.swift
@@ -1,20 +1,0 @@
-//
-//  CalculatorSymbol.swift
-//  Haruby-iOS
-//
-//  Created by 이정동 on 10/16/24.
-//
-
-import Foundation
-
-struct CalculatorSymbol {
-    static let plus = "+"
-    static let minus = "-"
-    static let multiply = "×"
-    static let divide = "÷"
-    static let equal = "="
-    static let deleteAll = "AC"
-    static let deleteLast = "Delete"
-    static let doubleZero = "00"
-    static let tripleZero = "000"
-}
