--- Haruby-iOS/Sources/Utility/Constant/Fonts.swift
@@ -1,8 +1,0 @@
-//
-//  Fonts.swift
-//  Haruby-iOS
-//
-//  Created by namdghyun on 10/10/24.
-//
-
-import Foundation
